updated output to look nicer
